Update Export Index with Workflow Type list
Update the Export Index action to retreieve a list of workflow types based
on the OrchardCore.Workflows WorkflowType Index action.

Update Index view to list out the workflow types, add an export button
which will point to the export route and update the links for properties
and edit so that they point to the right workflows actions.